build slider

+++ app/views/backend/articles/index.blade.php
@@ -2,7 +2,7 @@
 <div class="row"
      <div class="col-lg-12">
             <ol class="breadcrumb">
-              <li><a href=""><i class="fa fa-dashboard"></i>{{trans('common.menu.dashboard')}}</a></li>
+              <li><a href="{{Request::root()}}/backend"><i class="fa fa-dashboard"></i>{{trans('common.menu.dashboard')}}</a></li>
               <li class="active"><a href="{{Request::root()}}/backend/article"><i class="fa fa-desktop"></i> {{trans('common.table.article')}} </a></li>            
             </ol>
     </div>   
@@ -17,7 +17,7 @@
                 </div>
                 <div class="panel panel-success">
                     <div class="panel-heading"> 
-                     
+                        {{trans('titlepage.title.list_article')}}
                     </div>
                     <div class="panel-body">                        
                             <div class="col-lg-12">
@@ -56,8 +56,8 @@
                                         <td><a href="{{Request::root()}}/backend/article/view/{{$article->id}}">{{$article->title}}</a></td>
                                         <td>{{$article->status}}</td>
                                         <td>{{$article->create_by}}</td>
-                                        <td><a href="{{Request::root()}}/backend/article/update/{{$article->id}}"><span class="label label-primary">{{trans('common.button.update')}}</span></a>
-                                            <a href="{{Request::root()}}/backend/article/delete/{{$article->id}}"><span class="label label-danger">{{trans('common.button.delete')}}</span></a></td>
+                                        <td><a href="{{Request::root()}}/backend/article/update/{{$article->id}}" ><span class="label label-primary">{{trans('common.button.update')}}</span></a>
+                                            <a href="{{Request::root()}}/backend/article/delete/{{$article->id}}" onclick="return confirm('{{trans("messages.cf_delete")}}');"><span class="label label-danger">{{trans('common.button.delete')}}</span></a></td>
                                       </tr>                                      
                                       @endforeach
                                     </tbody>
